requirements.txt
python-telegram-bot==20.5
fastapi==0.115.0
uvicorn==0.30.0